refac(mangrove.js) rename and moved methods
part of #252
