+++ packages/mangrove.js/src/market.ts
@@ -5,8 +5,7 @@ import MgvToken from "./mgvtoken";
 import Semibook from "./semibook";
 import { Bigish, typechain } from "./types";
 import { Deferred } from "./util";
-import { logger } from "./util/logger";
-import TradeManagement from "./util/tradeManagement";
+import Trade from "./util/trade";
 
 let canConstructMarket = false;
 
@@ -205,7 +204,7 @@ class Market {
   #asksSemibook: Semibook;
   #bidsSemibook: Semibook;
   #initClosure?: () => Promise<void>;
-  tradeManagement: TradeManagement = new TradeManagement();
+  trade: Trade = new Trade();
   tradeEventManagement: TradeEventManagement = new TradeEventManagement();
 
   static async connect(params: {
@@ -276,8 +275,8 @@ class Market {
       chunkSize: opts.chunkSize,
       desiredPrice: opts.desiredPrice,
       desiredVolume:
-        (ba === "asks" && this.tradeManagement.getIsVolumeDesiredForAsks(opts)) ||
-        (ba === "bids" && this.tradeManagement.getIsVolumeDesiredForBids(opts))
+        (ba === "asks" && Semibook.getIsVolumeDesiredForAsks(opts)) ||
+        (ba === "bids" && Semibook.getIsVolumeDesiredForBids(opts))
           ? semibookDesiredVolume
           : undefined,
     });
@@ -468,48 +467,7 @@ class Market {
     params: Market.TradeParams,
     overrides: ethers.Overrides = {}
   ): Promise<Market.OrderResult> {
-    const { wants, givesWithoutSlippage, gives, fillWants } =
-      this.tradeManagement.getParamsForBuy(params, this.base, this.quote);
-    if ("restingOrder" in params && params.restingOrder) {
-      const makerWants = wants;
-      const makerGives = this.quote.toUnits(givesWithoutSlippage);
-
-      return this.#restingOrder(
-        {
-          gives,
-          makerGives,
-          wants,
-          makerWants,
-          orderType: "buy",
-          fillWants,
-          params: params.restingOrder,
-        },
-        overrides
-      );
-    } else {
-      if ("offerId" in params && params.offerId) {
-        return this.#snipes(
-          {
-            targets: [
-              {
-                offerId: params.offerId,
-                gives: gives,
-                wants: wants,
-                gasLimit: null,
-              },
-            ],
-            fillWants: fillWants,
-            orderType: "buy",
-          },
-          overrides
-        );
-      } else {
-        return this.#marketOrder(
-          { gives, wants, orderType: "buy", fillWants },
-          overrides
-        );
-      }
-    }
+    return this.trade.buy(params, overrides, this)
   }
 
   /**
@@ -538,402 +496,9 @@ class Market {
     params: Market.TradeParams,
     overrides: ethers.Overrides = {}
   ): Promise<Market.OrderResult> {
-    const { gives, wants, wantsWithoutSlippage, fillWants } =
-      this.tradeManagement.getParamsForSell(params, this.base, this.quote);
-    if ("restingOrder" in params && params.restingOrder) {
-      const makerGives = gives;
-      const makerWants = this.quote.toUnits(wantsWithoutSlippage);
-      return this.#restingOrder(
-        {
-          gives,
-          makerGives,
-          wants,
-          makerWants,
-          orderType: "sell",
-          fillWants,
-          params: params.restingOrder,
-        },
-        overrides
-      );
-    } else {
-      if ("offerId" in params && params.offerId) {
-        return this.#snipes(
-          {
-            targets: [
-              {
-                offerId: params.offerId,
-                gives: wants,
-                wants: gives,
-                gasLimit: null,
-              },
-            ],
-            orderType: "sell",
-            fillWants: fillWants,
-          },
-          overrides
-        );
-      } else {
-        return this.#marketOrder(
-          { wants, gives, orderType: "sell", fillWants },
-          overrides
-        );
-      }
-    }
-  }
+    return this.trade.sell(params, overrides, this)
+  } 
 
-  #resultOfEvent(
-    evt: ethers.Event,
-    got_bq: "base" | "quote",
-    gave_bq: "base" | "quote",
-    fillWants: boolean,
-    takerWants: ethers.BigNumber,
-    takerGives: ethers.BigNumber,
-    result: Market.OrderResult
-  ): Market.OrderResult {
-    return this.#resultOfEventCore(
-      evt,
-      got_bq,
-      gave_bq,
-      this.tradeEventManagement.partialFill(fillWants, takerWants, takerGives),
-      result
-    );
-  }
-
-  #resultOfEventCore(
-    evt: ethers.Event,
-    got_bq: "base" | "quote",
-    gave_bq: "base" | "quote",
-    partialFillFunc: (takerGot: BigNumber, takerGave: BigNumber) => boolean,
-    result: Market.OrderResult
-  ): Market.OrderResult {
-    const got = this[got_bq];
-    const gave = this[gave_bq];
-    switch (evt.event) {
-      case "OrderComplete": {
-        result.summary = this.tradeEventManagement.createSummaryForOrderComplete(
-          evt,
-          got,
-          gave,
-          partialFillFunc
-        );
-        return result;
-      }
-      case "OfferSuccess": {
-        result.successes.push(this.tradeEventManagement.createSuccess(evt, got, gave));
-        return result;
-      }
-      case "OfferFail": {
-        result.tradeFailures.push(
-          this.tradeEventManagement.createTradeFailure(evt, got, gave)
-        );
-        return result;
-      }
-      case "PosthookFail": {
-        result.posthookFailures.push(
-          this.tradeEventManagement.createPosthookFailure(evt)
-        );
-        return result;
-      }
-      case "OrderSummary": {
-        result.summary = this.tradeEventManagement.createSummaryForOrderSummary(
-          evt,
-          got,
-          gave,
-          partialFillFunc
-        );
-        return result;
-      }
-      default: {
-        return result;
-      }
-    }
-  }
-  /**
-   * Low level Mangrove market order.
-   * If `orderType` is `"buy"`, the base/quote market will be used,
-   *
-   * If `orderType` is `"sell"`, the quote/base market will be used,
-   *
-   * `fillWants` defines whether the market order stops immediately once `wants` tokens have been purchased or whether it tries to keep going until `gives` tokens have been spent.
-   *
-   * In addition, `slippage` defines an allowed slippage in % of the amount of quote token.
-   *
-   * Returns a promise for market order result after 1 confirmation.
-   * Will throw on same conditions as ethers.js `transaction.wait`.
-   */
-  async #marketOrder(
-    {
-      wants,
-      gives,
-      orderType,
-      fillWants,
-    }: {
-      wants: ethers.BigNumber;
-      gives: ethers.BigNumber;
-      orderType: "buy" | "sell";
-      fillWants: boolean;
-    },
-    overrides: ethers.Overrides
-  ): Promise<Market.OrderResult> {
-    const [outboundTkn, inboundTkn] =
-      orderType === "buy" ? [this.base, this.quote] : [this.quote, this.base];
-
-    // user defined gasLimit overrides estimates
-    if (!overrides.gasLimit) {
-      overrides.gasLimit = await this.estimateGas(orderType, wants);
-    }
-
-    logger.debug("Creating market order", {
-      contextInfo: "market.marketOrder",
-      data: {
-        outboundTkn: outboundTkn.name,
-        inboundTkn: inboundTkn.name,
-        fillWants: fillWants,
-        wants: wants.toString(),
-        gives: gives.toString(),
-        orderType: orderType,
-        gasLimit: overrides.gasLimit?.toString(),
-      },
-    });
-    const response = await this.mgv.contract.marketOrder(
-      outboundTkn.address,
-      inboundTkn.address,
-      wants,
-      gives,
-      fillWants,
-      overrides
-    );
-    const receipt = await response.wait();
-
-    //last OrderComplete is ours!
-    logger.debug("Market order raw receipt", {
-      contextInfo: "market.marketOrder",
-      data: { receipt: receipt },
-    });
-    const result: Market.OrderResult = this.createOrderResultFromReceipt(
-      receipt,
-      orderType,
-      fillWants,
-      wants,
-      gives,
-      this.mgv._address
-    );
-    if (!result.summary) {
-      throw Error("market order went wrong");
-    }
-    return result;
-  }
-
-  async #restingOrder(
-    {
-      wants,
-      makerWants,
-      gives,
-      makerGives,
-      orderType,
-      fillWants,
-      params,
-    }: {
-      wants: ethers.BigNumber;
-      makerWants: ethers.BigNumber;
-      gives: ethers.BigNumber;
-      makerGives: ethers.BigNumber;
-      orderType: "buy" | "sell";
-      fillWants: boolean;
-      params: Market.RestingOrderParams;
-    },
-    overrides: ethers.Overrides
-  ): Promise<Market.OrderResult> {
-    const overrides_ = {
-      ...overrides,
-      value: this.mgv.toUnits(params.provision, 18),
-    };
-
-    // user defined gasLimit overrides estimates
-    overrides_.gasLimit = overrides_.gasLimit
-      ? overrides_.gasLimit
-      : await this.estimateGas(orderType, wants);
-
-    const response = await this.mgv.orderContract.take(
-      {
-        base: this.base.address,
-        quote: this.quote.address,
-        partialFillNotAllowed: params.partialFillNotAllowed
-          ? params.partialFillNotAllowed
-          : false,
-        selling: orderType === "sell",
-        wants: wants,
-        makerWants: makerWants,
-        gives: gives,
-        makerGives: makerGives,
-        restingOrder: true,
-        retryNumber: params.retryNumber ? params.retryNumber : 0,
-        gasForMarketOrder: params.gasForMarketOrder
-          ? params.gasForMarketOrder
-          : 0,
-        blocksToLiveForRestingOrder: params.blocksToLiveForRestingOrder
-          ? params.blocksToLiveForRestingOrder
-          : 0,
-      },
-      overrides_
-    );
-    const receipt = await response.wait();
-
-    //last OrderComplete is ours!
-    logger.debug("Resting order raw receipt", {
-      contextInfo: "market.restingOrder",
-      data: { receipt: receipt },
-    });
-
-    const result: Market.OrderResult = this.createOrderResultFromReceipt(
-      receipt,
-      orderType,
-      fillWants,
-      wants,
-      gives,
-      this.mgv.orderContract.address
-    );
-    if (!result.summary) {
-      throw Error("resting order went wrong");
-    }
-    // if resting order was not posted, result.summary is still undefined.
-    return result;
-  }
-
-  private createOrderResultFromReceipt(
-    receipt: ethers.ethers.ContractReceipt,
-    orderType: string,
-    fillWants: boolean,
-    wants: ethers.ethers.BigNumber,
-    gives: ethers.ethers.BigNumber,
-    address: string
-  ) {
-    let result: Market.OrderResult = {
-      txReceipt: receipt,
-      summary: undefined,
-      successes: [],
-      tradeFailures: [],
-      posthookFailures: [],
-    };
-    const got_bq = orderType === "buy" ? "base" : "quote";
-    const gave_bq = orderType === "buy" ? "quote" : "base";
-    for (const evt of receipt.events) {
-      if (
-        evt.address === address &&
-        (!evt.args.taker || receipt.from === evt.args.taker)
-      ) {
-        result = this.#resultOfEvent(
-          evt,
-          got_bq,
-          gave_bq,
-          fillWants,
-          wants,
-          gives,
-          result
-        );
-      }
-    }
-    return result;
-  }
-
-  /**
-   * Low level sniping of `targets`.
-   *
-   * If `orderType` is `"buy"`, the base/quote market will be used,
-   * If `orderType` is `"sell"`, the quote/base market will be used,
-   *
-   * `fillWants` defines whether the market order stops immediately once `wants` tokens have been purchased or whether it tries to keep going until `gives` tokens have been spent.
-   *
-   * Returns a promise for snipes result after 1 confirmation.
-   * Will throw on same conditions as ethers.js `transaction.wait`.
-   */
-  async #snipes(
-    {
-      targets,
-      orderType,
-      fillWants,
-    }: {
-      targets: {
-        offerId: ethers.BigNumberish;
-        wants: ethers.BigNumber;
-        gives: ethers.BigNumber;
-        gasLimit?: ethers.BigNumber;
-      }[];
-      orderType: "buy" | "sell";
-      fillWants: boolean;
-    },
-    overrides: ethers.Overrides
-  ): Promise<Market.OrderResult> {
-    const [outboundTkn, inboundTkn] =
-      orderType === "buy" ? [this.base, this.quote] : [this.quote, this.base];
-
-    logger.debug("Creating snipes", {
-      contextInfo: "market.snipes",
-      data: {
-        outboundTkn: outboundTkn.name,
-        inboundTkn: inboundTkn.name,
-        fillWants: fillWants,
-      },
-    });
-
-    // user defined gasLimit overrides estimates
-    const _targets = targets.map<
-      [
-        ethers.BigNumberish | Promise<ethers.BigNumberish>,
-        ethers.BigNumberish | Promise<ethers.BigNumberish>,
-        ethers.BigNumberish | Promise<ethers.BigNumberish>,
-        ethers.BigNumberish | Promise<ethers.BigNumberish>
-      ]
-    >((t) => [
-      t.offerId,
-      t.wants,
-      t.gives,
-      t.gasLimit ?? overrides.gasLimit ?? this.estimateGas(orderType, t.wants),
-    ]);
-
-    const response = await this.mgv.contract.snipes(
-      outboundTkn.address,
-      inboundTkn.address,
-      _targets,
-      fillWants,
-      overrides
-    );
-
-    const receipt = await response.wait();
-
-    let result: Market.OrderResult = {
-      txReceipt: receipt,
-      summary: undefined,
-      successes: [],
-      tradeFailures: [],
-      posthookFailures: [],
-    };
-    //last OrderComplete is ours!
-    logger.debug("Snipes raw receipt", {
-      contextInfo: "market.snipes",
-      data: { receipt: receipt },
-    });
-    const got_bq = orderType === "buy" ? "base" : "quote";
-    const gave_bq = orderType === "buy" ? "quote" : "base";
-    for (const evt of receipt.events) {
-      if (
-        evt.address === this.mgv._address &&
-        (!evt.args.taker || receipt.from === evt.args.taker)
-      ) {
-        result = this.#resultOfEventCore(
-          evt,
-          got_bq,
-          gave_bq,
-          () => false,
-          result
-        );
-      }
-    }
-    if (!result.summary) {
-      throw Error("snipes went wrong");
-    }
-    return result;
-  }
 
   async estimateGas(bs: "buy" | "sell", volume: BigNumber): Promise<BigNumber> {
     const semibook = bs === "buy" ? this.#asksSemibook : this.#bidsSemibook;
